+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.1.0 / 15/06/2022
 - resolução do desafio de comparação usando operadores e sintaxe
 - resolução da atividade verifica palindromo (aula sobre variáveis e tipos)
+- resolução da atividade substituir numeros pares
 
 ## 1.0.0 / 14/06/2022
 - criada a estrutura do projeto